requirements.txt
gradio>=4.0.0
numpy>=1.21.0
pillow>=8.0.0
matplotlib>=3.5.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.9.0
transformers>=4.20.0 